requirements.txt
pymupdf
pdfplumber
python-docx
